+++ API/Dtos/Lesson/UpdateLesssonDto.cs
@@ -13,7 +13,7 @@ namespace API.Dtos.Lesson
         public string UrlPractice { get; set; }
         public int Number { get; set; } = -1;
         public int Importance { get; set; } = -1;
-        public int TestScore { get; set; } = -1;
+        public float TestScore { get; set; } = -1;
         public int IsTheoryCompleted { get; set; } = -1;
         public int IsPracticeCompleted { get; set; } = -1;
     }
