+++ API/Entities/Lesson.cs
@@ -9,7 +9,7 @@ namespace API.Entities
         public string UrlPractice { get; set; }
         public int Number { get; set; }
         public int Importance { get; set; } = 0;
-        public int TestScore { get; set; } = -1;
+        public float TestScore { get; set; } = -1;
         public bool IsTheoryCompleted { get; set; } = false;
         public bool IsPracticeCompleted { get; set; } = false;
         public List<SectionLesson> SectionLessons { get; set; }
